Move `BuildManager` and `Build` to their own module and make them public
